refactor: simplify thread deletion logic in messageDelete event

+++ src/events/messageDelete.js
@@ -37,12 +37,9 @@ module.exports = {
 
 			//logger.DBG(`Message deleted in [${message.channel.name}] by ${message.author.tag}: ${message.content}`);
 
-			if (message.channelId === channels.CREACIONES) {
-				logger.DBG(message.hasThread)
-				if (message.hasThread) {
-					const thread = message.thread ?? await message.channel.threads.fetch(message.id);
-					await thread.delete();
-				}
+			if (message.channelId === channels.CREACIONES && message.hasThread) {
+				const thread = message.thread ?? await message.channel.threads.fetch(message.id);
+				await thread.delete();
 			}
 		} catch (error) {
 			logger.ERR(error);
